# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: MacQuiz v2 API
  description: Quiz platform for admins, teachers, and students.
  version: 0.8.0
servers:
  - url: /
paths:
...
  /api/v1/quizzes/{id}/assignments:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: listAssignments
      summary: The quiz's current audience (owner)
      responses:
        "200":
          description: The assigned students.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: setAssignments
      summary: Replace the quiz's audience (owner)
      description: >-
        Group ids are expanded to individual student rows at assignment
        time, so later group edits never silently revoke an already-assigned
        quiz. Atomic - one bad id changes nothing. Allowed while draft or
        scheduled; a live quiz's audience is frozen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAssignmentsRequest"
      responses:
        "200":
          description: The resulting audience, group-expanded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/NotEditable"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /api/v1/quizzes/{id}/release-results:
    post:
      operationId: releaseResults
      summary: Release results to students (owner)
      description: >-
        Stamps results_released_at, the fact every student results read
        gates on. Requires the quiz's window to have ended (409
        QUIZ_NOT_CLOSED otherwise); releasing an already-released quiz is an
        idempotent no-op. Quizzes published with release_policy auto are
        released by the worker at close without this call.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The quiz with results_released_at set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/QuizNotClosed"
...
components:
  responses:
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Body-level validation failed; see fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The policy denies this action for the caller's role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >-
        No such resource. Also the answer for resources the caller is not
        assigned to, so existence is never leaked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotEditable:
      description: >-
        QUIZ_NOT_EDITABLE - the quiz has been published; its question set is
        an immutable snapshot and drafts-only operations are refused.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    QuizNotClosed:
      description: >-
        QUIZ_NOT_CLOSED - results can be released once the quiz's window has
        ended.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            Stable machine-readable code from the docs/04-api.md vocabulary,
            e.g. INVALID_CREDENTIALS, UNAUTHENTICATED, VALIDATION_FAILED,
            RATE_LIMITED, PASSWORD_CHANGE_REQUIRED, NOT_FOUND.
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
          description: Per-field messages for VALIDATION_FAILED.
...
    Quiz:
      type: object
      required:
        [
          id,
          owner_id,
          title,
          status,
          starts_at,
          ends_at,
          duration_sec,
          max_attempts,
          shuffle_questions,
          published_at,
          version,
          created_at,
          question_count,
          release_policy,
          results_released_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, scheduled, live, closed, archived]
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        duration_sec:
          type: integer
          nullable: true
        max_attempts:
          type: integer
        shuffle_questions:
          type: boolean
        published_at:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        question_count:
          type: integer
        release_policy:
          type: string
          enum: [auto, manual]
          description: >-
            When scores become visible to students - auto releases in the
            worker pass that grades the closed quiz, manual waits for
            POST /quizzes/{id}/release-results.
        results_released_at:
          type: string
          format: date-time
          nullable: true
          description: Null means scores and the answer key are withheld.
...
    QuizResponse:
      type: object
      required: [quiz]
      properties:
        quiz:
          $ref: "#/components/schemas/Quiz"
...
    SetAssignmentsRequest:
      type: object
      properties:
        student_ids:
          type: array
          items:
            type: string
            format: uuid
        group_ids:
          type: array
          description: Expanded to individual student rows at assignment time.
          items:
            type: string
            format: uuid
    AssignedStudent:
      type: object
      required: [id, full_name, email]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
    AssignmentsResponse:
      type: object
      required: [students]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/AssignedStudent"
...